pyproject.toml
[project]
name = "ai-document-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit",
    "google-generativeai",
    "deepeval",
    "pytest",
    "pytest-html",
    "python-dotenv"
]
